split anidb.py, save xml to database
