+++ fadbs/util/__init__.py
@@ -1,3 +1,4 @@
 """ Utilities for the rest of FADBS """
-from .anidb import AnidbParser
-from .anidb import AnidbSearch
+from . import api_anidb
+from .anidb_search import AnidbSearch
+from .anidb_parse import AnidbParser
